Rename `app create` to `app template`, add `app create` & `app remove`
