+++ src/cli/app/index.ts
@@ -3,6 +3,8 @@ import * as generate from './generate.js';
 import * as install from './install.js';
 import * as list from './list.js';
 import * as permission from './permission.js';
+import * as remove from './remove.js';
+import * as template from './template.js';
 import * as token from './token.js';
 import * as tunnel from './tunnel.js';
 import * as uninstall from './uninstall.js';
@@ -16,6 +18,8 @@ export default function (_: any) {
     tunnel,
     token,
     permission,
+    template,
+    remove,
 
     // no auth needed
     generate,
